+++ src-springboot/src/main/java/com/arturviader/pelisbdapi/dto/MovieTMDB.java
@@ -4,12 +4,12 @@ import lombok.Data;
 
 @Data
 public class MovieTMDB {
-    private int id;
+    private Long id;
     private String title;
     private String release_date;
     private String overview;
 
-    public MovieTMDB(int id, String title, String release_date, String overview, String poster_path) {
+    public MovieTMDB(Long id, String title, String release_date, String overview, String poster_path) {
         this.id = id;
         this.title = title;
         this.release_date = release_date;
@@ -19,8 +19,8 @@ public class MovieTMDB {
 
     private String poster_path;
 
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
